split actions

+++ backend/composer/run.sh
@@ -3,6 +3,16 @@
 # determine this files directory
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-rm -rf $DIR/../vendor
-rm -f $DIR/../composer.lock
-docker run --rm -v "$DIR/../":/app opinephp/backendcomposer install
+MODE="install"
+if [ $# -eq 1 ]
+then
+    MODE=$1
+fi
+
+if [ $MODE -eq "install"]
+then
+    rm -rf $DIR/../vendor
+    rm -f $DIR/../composer.lock
+fi
+
+docker run --rm -v "$DIR/../":/app opinephp/backendcomposer $MODE
